+++ java/socket/src/msgs/FloatSeq.java
@@ -62,7 +62,7 @@ public class FloatSeq extends SocketMsg {
 		readSequence(_in);
 	}
 
-	public void deserialize(long _id, Hash _hash, int _type, long _size, DataInputStream _in) throws IOException {
+	public void deserialize(long _id, Hash _hash, int _type, int _size, DataInputStream _in) throws IOException {
 		super.deserialize(_id, _hash, _type, _size, _in);
 		readSequence(_in);
 	}
