+++ java/socket/src/msgs/Int.java
@@ -57,7 +57,7 @@ public class Int extends SocketMsg {
 		value = _in.readInt();
 	}
 	
-	public void deserialize(long _id, Hash _hash, int _type, long _size, DataInputStream _in) throws IOException {
+	public void deserialize(long _id, Hash _hash, int _type, int _size, DataInputStream _in) throws IOException {
 		super.deserialize(_id, _hash, _type, _size, _in);
 		value = _in.readInt();
 		System.out.println(value);
